+++ qtrmrs/apps/users/views.py
@@ -1,4 +1,5 @@
-from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse
 from django_ratelimit.decorators import ratelimit
 from django.contrib.auth import login, logout
 from django.contrib.auth.decorators import login_required